+++ sample/src/main/java/com/arttttt/simplemvi/sample/counter/CounterStore.kt
@@ -5,8 +5,8 @@ import com.arttttt.simplemvi.store.Store
 import com.arttttt.simplemvi.actor.ActorScope
 import com.arttttt.simplemvi.actor.dsl.ActorBuilder
 import com.arttttt.simplemvi.logging.loggingActor
-import com.arttttt.simplemvi.utils.actorDsl
-import com.arttttt.simplemvi.utils.createStore
+import com.arttttt.simplemvi.actor.dsl.actorDsl
+import com.arttttt.simplemvi.store.createStore
 import kotlin.coroutines.CoroutineContext
 
 class CounterStore(
