+++ sample/src/main/java/com/arttttt/simplemvi/sample/notes/domain/stores/NotesStore.kt
@@ -5,8 +5,8 @@ import com.arttttt.simplemvi.logging.loggingActor
 import com.arttttt.simplemvi.sample.notes.domain.models.Note
 import com.arttttt.simplemvi.sample.notes.domain.repository.NotesRepository
 import com.arttttt.simplemvi.store.Store
-import com.arttttt.simplemvi.utils.actorDsl
-import com.arttttt.simplemvi.utils.createStore
+import com.arttttt.simplemvi.actor.dsl.actorDsl
+import com.arttttt.simplemvi.store.createStore
 import kotlinx.coroutines.launch
 import kotlin.coroutines.CoroutineContext
 import kotlin.uuid.Uuid
