+++ sample/src/main/java/com/arttttt/simplemvi/sample/notes/presentation/NotesViewModel.kt
@@ -6,8 +6,8 @@ import com.arttttt.simplemvi.sample.notes.domain.repository.NotesRepository
 import com.arttttt.simplemvi.sample.notes.domain.stores.NotesStore
 import com.arttttt.simplemvi.sample.notes.presentation.models.NoteListItem
 import com.arttttt.simplemvi.sample.ui.ListItem
-import com.arttttt.simplemvi.utils.plus
-import com.arttttt.simplemvi.utils.state
+import com.arttttt.simplemvi.store.plus
+import com.arttttt.simplemvi.store.state
 import com.arttttt.simplemvi.viewmodel.attachStore
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.SharingStarted
