+++ sample/src/main/java/com/arttttt/simplemvi/sample/timer/TimerContent.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.arttttt.simplemvi.sample.timer.store.TimerStore
-import com.arttttt.simplemvi.utils.plus
+import com.arttttt.simplemvi.store.plus
 
 @Composable
 fun TimerContent() {
