+++ simplemvi/src/androidMain/kotlin/com/arttttt/simplemvi/utils/MainThreadAssert.android.kt
@@ -1,4 +1,4 @@
-package com.arttttt.simplemvi.utils.mainthread
+package com.arttttt.simplemvi.utils
 
 import android.os.Looper
