+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/actor/DefaultActor.kt
@@ -1,7 +1,7 @@
 package com.arttttt.simplemvi.actor
 
-import com.arttttt.simplemvi.utils.mainthread.MainThread
-import com.arttttt.simplemvi.utils.mainthread.assertOnMainThread
+import com.arttttt.simplemvi.utils.MainThread
+import com.arttttt.simplemvi.utils.assertOnMainThread
 import kotlinx.atomicfu.atomic
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.cancel
