+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/utils/MainThread.kt
@@ -1,4 +1,4 @@
-package com.arttttt.simplemvi.utils.mainthread
+package com.arttttt.simplemvi.utils
 
 @Retention(AnnotationRetention.SOURCE)
 annotation class MainThread
