+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/utils/MainThreadAssert.common.kt
@@ -1,4 +1,4 @@
-package com.arttttt.simplemvi.utils.mainthread
+package com.arttttt.simplemvi.utils
 
 @MainThread
 fun assertOnMainThread() {
