+++ simplemvi/src/iosMain/kotlin/com/arttttt/simplemvi/utils/MainThreadAssert.ios.kt
@@ -1,4 +1,4 @@
-package com.arttttt.simplemvi.utils.mainthread
+package com.arttttt.simplemvi.utils
 
 import platform.Foundation.NSThread
